+++ src/main/java/com/lothrazar/cyclicmagic/component/pump/item/BlockItemPump.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.component.pumpitem;
+package com.lothrazar.cyclicmagic.component.pump.item;
 import com.lothrazar.cyclicmagic.IHasRecipe;
 import com.lothrazar.cyclicmagic.block.base.BlockBaseFacingOmni;
 import com.lothrazar.cyclicmagic.gui.ForgeGuiHandler;
