+++ src/main/java/com/lothrazar/cyclicmagic/component/pump/item/ContainerItemPump.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.component.pumpitem;
+package com.lothrazar.cyclicmagic.component.pump.item;
 import com.lothrazar.cyclicmagic.data.Const;
 import com.lothrazar.cyclicmagic.gui.base.ContainerBaseMachine;
 import com.lothrazar.cyclicmagic.gui.slot.SlotSingleStack;
