+++ src/main/java/com/lothrazar/cyclicmagic/component/pump/item/TileEntityItemPump.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.component.pumpitem;
+package com.lothrazar.cyclicmagic.component.pump.item;
 import java.util.List;
 import com.lothrazar.cyclicmagic.block.base.TileEntityBaseMachineInvo;
 import com.lothrazar.cyclicmagic.component.cable.TileEntityCableBase;
