+++ src/main/java/com/lothrazar/cyclicmagic/module/BlockUtilityModule.java
@@ -31,12 +31,12 @@ import com.lothrazar.cyclicmagic.component.itemsort.TileEntityItemCableSort;
 import com.lothrazar.cyclicmagic.component.library.BlockLibrary;
 import com.lothrazar.cyclicmagic.component.library.BlockLibraryController;
 import com.lothrazar.cyclicmagic.component.library.TileEntityLibrary;
-import com.lothrazar.cyclicmagic.component.pumpenergy.BlockEnergyPump;
-import com.lothrazar.cyclicmagic.component.pumpenergy.TileEntityEnergyPump;
-import com.lothrazar.cyclicmagic.component.pumpfluid.BlockFluidPump;
-import com.lothrazar.cyclicmagic.component.pumpfluid.TileEntityFluidPump;
-import com.lothrazar.cyclicmagic.component.pumpitem.BlockItemPump;
-import com.lothrazar.cyclicmagic.component.pumpitem.TileEntityItemPump;
+import com.lothrazar.cyclicmagic.component.pump.energy.BlockEnergyPump;
+import com.lothrazar.cyclicmagic.component.pump.energy.TileEntityEnergyPump;
+import com.lothrazar.cyclicmagic.component.pump.fluid.BlockFluidPump;
+import com.lothrazar.cyclicmagic.component.pump.fluid.TileEntityFluidPump;
+import com.lothrazar.cyclicmagic.component.pump.item.BlockItemPump;
+import com.lothrazar.cyclicmagic.component.pump.item.TileEntityItemPump;
 import com.lothrazar.cyclicmagic.component.scaffold.BlockScaffolding;
 import com.lothrazar.cyclicmagic.component.scaffold.BlockScaffoldingReplace;
 import com.lothrazar.cyclicmagic.component.scaffold.BlockScaffoldingResponsive;
